split util into .cc file
